request_service/requirements.txt
aio_pika==9.0.5
aiohttp==3.8.4
PyYAML==6.0
pydantic==1.10.9